% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{annualDomesticHarvest}
\alias{annualDomesticHarvest}
\title{Annual Domestic Harvest}
\usage{
annualDomesticHarvest(years, onlyvar = FALSE)
}
\arguments{
\item{years}{years to calculate statistics for}

\item{onlyvar}{If true, only the variable 5 will be returned.
If false, other calculations needed to produce variable 5 will be returned}
}
\value{
Depends on onlyvar, either data frame of intermediate calculations or a vector
of values for Annual Domestic Harvest
}
\description{
Various statistics and intermediate calculations to solve for
Variable 5, Annual Domestic Harvest (H)
}

