% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericFlow.R
\name{calcP_EX}
\alias{calcP_EX}
\title{Statistics for variable 4}
\usage{
calcP_EX(years = 1990:2020, var = FALSE)
}
\arguments{
\item{years}{years to calculate}

\item{var}{If true, return only variable 4. If false, return intermediate statistics needed for variable 4}
}
\description{
Statistics for variable 4
}

