% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericFlow.R
\name{calcP_IM}
\alias{calcP_IM}
\title{Statistics for variable 3}
\usage{
calcP_IM(years = 1990:2020, var = FALSE)
}
\arguments{
\item{years}{years to calculate for}

\item{var}{if true, returns variable 3. if false, returns statistics necessary to calculate variable 3}
}
\description{
Statistics for variable 3
}

