% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swp-carbon.R
\name{calculateswp}
\alias{calculateswp}
\title{Title}
\usage{
calculateswp()
}
\value{
Data frame of Roundwood and solidwood production, imports and exports
}
\description{
Title
}
\examples{
calculateswp()
}

