% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{calcUSApaper}
\alias{calcUSApaper}
\title{Calculates paper statistics}
\usage{
calcUSApaper(years = yrs, woodToCarbon = 4.535925e-07,
  paperToCarbon = 3.9008955e-07)
}
\arguments{
\item{years}{years to calculate data for}
}
\value{
A data frame with necessary intermediate calculations to calculate carbon from paper.

Corresponds with necessary columns from the `USA` sheet in the WOODCARB spreadsheet
}
\description{
Uses anonymous functions to calculate columns from `USA` sheet in WOODCARB spreadsheet.
}

