% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papercarbon.R
\name{calcUSAproduction}
\alias{calcUSAproduction}
\title{Calculates paper statistics}
\usage{
calcUSAproduction()
}
\value{
a data frame with necessary intermediate calculations
}
\description{
Calculates paper statistics
}
\examples{
calcUSAproduction()
}

