% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{carbonfromdumps}
\alias{carbonfromdumps}
\title{Calculates carbon in landfills and dumps.}
\usage{
carbonfromdumps(years, approach = c("Production", "Stock Change"))
}
\arguments{
\item{years}{years to calculate values for}

\item{approach}{approach to use}
}
\value{
total carbon from solid waste disposal sites
}
\description{
Calculates carbon in landfills and dumps.
}

