% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{finalCarbonContribution}
\alias{finalCarbonContribution}
\title{Estimates U.S. HWP contribution to annual greenhouse gas removals.}
\usage{
finalCarbonContribution(Years = 1990:2015, approach = c("Production",
  "Stock Change", "Atmospheric Flow"), decaydistribution = c("Exponential",
  "K=2"), plot = FALSE, halflives = halfLives, paperHL = 2.53087281800454,
  fsp = fracstrpanels, fnsp = fracnonstrpanels, fsawn = fracsawnwood)
}
\arguments{
\item{Years}{years to return carbon totals for}

\item{approach}{The approach used to calculate carbon contribution.
"Production" approach corresponds to `06 IPCC Tables`$R9 in the WOODCARB II spreadsheet
"Stock Change" approach corresponds to `06 IPCC Tables`$P9 in the WOODCARB II spreadsheet
"Atmospheric Flow" approach corresponds to `06 IPCC Tables`$Q9 in the WOODCARB II spreadsheet}

\item{decaydistribution}{Type of decay method to use}

\item{plot}{if true, returns a simple plot of time vs. carbon contribution.}

\item{halflives}{data frame of half lives to use. Must have a column associated with each
one of 13 end uses and half life values for 1900 until the maximum year of interest.}

\item{paperHL}{half life value for paper}

\item{fsp}{Fraction of structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fnsp}{Fraction of non-structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fsawn}{Fraction of sawnwood products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}
}
\value{
A vector of carbon contributions for all Years. Plot returned is optional
All units returned are in Thousand metric tons CO2 Equivalent.
}
\description{
Uses formulas described in WOODCARB II paper to calculate contribution contribution
with the HWP variables.
}
\details{
Decay type and half life options allow for analysis of contribution levels under
different circumstances.

The Production Approach focuses on estimating annual carbon stock change in HWP and
forests where the carbon is from trees harvested in the reporting country.

The Stock Change Approach focuses on estimating annual carbon stock change in HWP and
forests in a country regardless of wood origin.

The Atmospheric Flow Approach focuses on estimating annual carbon fluxes between
the atmosphere and forests/HWPs within a country.
}
\examples{
finalCarbonContribution()
finalCarbonContribution(approach = "Stock Change")
finalCarbonContribution(approach = "Production",
                         decaydistribution ="Exponential")
}

