% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fracnonstrpanels}
\alias{fracnonstrpanels}
\title{Fractions of Non-Structural Panel production used in various end uses}
\format{A data frame with 151 observations of 16 variables:
\describe{
\item{House.SingFam}{Fraction of non-structural panel production in Single Family Homes}
\item{House.Multifam}{Fraction of non-structural panel production in Multi Family Homes}
\item{House.MobHom}{Fraction of non-structural panel production in Mobile Homes}
\item{Res.Upkeep}{Fraction of non-structural panel production in Total Residential Upkeep and Improvement}
\item{New.Nonres.AllRR}{Fraction of non-structural panel production in All Railroads}
\item{New.Nonres.Rties}{Fraction of non-structural panel production in Railroad Ties}
\item{New.Nonres.Rcar.Repair}{Fraction of non-structural panel production in Railcar Repair}
\item{Manu.HouseFurniture}{Fraction of non-structural panel production in Household Furniture}
\item{Manu.CommFurniture}{Fraction of non-structural panel production in Commercial Furniture}
\item{Manu.OtherProducts}{Fraction of non-structural panel production in Manufacturing and Other Products}
\item{Shipping.Tot}{Fraction of non-structural panel production in Shipping}
\item{Other.Uses.Tot}{Fraction of non-structural panel production in Other Uses}
\item{Other.Industrial.Tot}{Fraction of non-structural panel production in Uses For Other Industrial Products}
}}
\source{
Check spreadsheet - values entered in
}
\usage{
fracnonstrpanels
}
\description{
All units are percentages and all vectors are numeric.
}
\keyword{datasets}

