% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fracsawnwood}
\alias{fracsawnwood}
\title{Fractions of Sawnwood production used in various end uses}
\format{A data frame with 151 observations of 13 variables:
\describe{
\item{House.SingFam}{Fraction of sawnwood production in Single Family Homes}
\item{House.Multifam}{Fraction of sawnwood production in Multi Family Homes}
\item{House.MobHom}{Fraction of sawnwood production in Mobile Homes}
\item{Res.Upkeep}{Fraction of sawnwood production in Total Residential Upkeep and Improvement}
\item{New.Nonres.AllRR}{Fraction of sawnwood production in All Railroads}
\item{New.Nonres.Rties}{Fraction of sawnwood production in Railroad Ties}
\item{New.Nonres.Rcar.Repair}{Fraction of sawnwood production in Railcar Repair}
\item{Manu.HouseFurniture}{Fraction of sawnwood production in Household Furniture}
\item{Manu.CommFurniture}{Fraction of sawnwood production in Commercial Furniture}
\item{Manu.OtherProducts}{Fraction of sawnwood production in Manufacturing and Other Products}
\item{Shipping.Tot}{Fraction of sawnwood production in Shipping}
\item{Other.Uses.Tot}{Fraction of sawnwood production in Other Uses}
\item{Other.Industrial.Tot}{Fraction of sawnwood production in Uses For Other Industrial Products}
}}
\source{
Check spreadsheet - values entered in
}
\usage{
fracsawnwood
}
\description{
All units are percentages and all vectors are numeric.
}
\keyword{datasets}

