% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fracstrpanels}
\alias{fracstrpanels}
\title{Fractions of Structual Panel production used in various end uses}
\format{A data frame with 151 observations of 16 variables.

\itemize{
\item \strong{House.SingFam} Single Family Homes
\item \strong{House.Multifam} Multi Family Homes
\item \strong{House.MobHom} Mobile Homes
\item \strong{Res.Upkeep} Total Residential Upkeep and Improvement
\item \strong{New.Nonres.AllRR} All Railroads
\item \strong{New.Nonres.Rties} Railroad Ties
\item \strong{New.Nonres.Rcar.Repair} Railcar Repair
\item \strong{Manu.HouseFurniture} Household Furniture
\item \strong{Manu.CommFurniture} Commercial Furniture
\item \strong{Manu.OtherProducts} Manufacturing - Other Products
\item \strong{Shipping.Tot} Shipping
\item \strong{Other.Uses.Tot} Other Uses
\item \strong{other.industrial.tot} Uses For Other Industrial Products
}}
\source{
Check spreadsheet - values entered in
}
\usage{
fracstrpanels
}
\description{
All units are percentages and all vectors are numeric.
}
\keyword{fracstrpanels}

