% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{hair1958}
\alias{hair1958}
\title{hair1958}
\format{A data frame with 58 observations on 23 variables}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
hair1958
}
\description{
Production, imports, exports, and new supply of lumber in US between 1799 and 1958
}
\keyword{hair1958}

