% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hair1963}
\alias{hair1963}
\title{Estimated production, net imports, and apparent consumption of timber products in the US, 1900-1962}
\format{A data frame with 85 observations on 38 variables:
\describe{
\item{Years}{Years}
\item{Dom.Prod.Tot}{Total Domestic Production}
\item{ApparentConsumption}{Apparent Consumption}
\item{Dom.Prod}{Domestic Production, Total Ind RW}
\item{Imports}{Net Imports, Total Ind RW}
\item{Prod.SW}{Softwood Production, Total Ind RW}
\item{Prod.HW}{Hardwood Production , Total Ind RW}
\item{LogChipImports.SW}{Softwood Log and Chip Imports, Total Ind RW}
\item{LogChipExports.SW}{Softwood Log and Chip Exports, Total Ind RW}
\item{LogChipImports.HW}{Hardwood Log and Chip Imports, Total Ind RW}
\item{LogChipExports.HW}{Hardwood Log and Chip Exports, Total Ind RW}
\item{SW.PctOfProd}{Softwoods as Percent of Production, Total Ind RW}
\item{Exports.SW.PctOfProd}{Softwood Exports as Percent of Production, Total Ind RW}
\item{Exports.HW.PctOfProd}{Hardwood Exports as Percent of Production, Total Ind RW}
\item{Imports.SW.PctOfConsump}{Softwood Imports as Percent of Production, Total Ind RW}
\item{Imports.HW.PctOfConsump}{Hardwood Imports as Percent of Production, Total Ind RW}
\item{SW.Lumber.PctOfTotLumbProd}{Softwood Lumber as Percent of Total Lumber Production, Total Ind RW}
\item{DV.Exports.Log.PctOfProd}{Dollar Value of Log Exports as Percent of Production, Total Ind RW}
\item{DV.Imports.Log.PctOfProd}{Dollar Value of Log Imports as Percent of Consumption, Total Ind RW}
\item{Estimated.Imports}{Estimated Imports, Total Ind RW}
\item{Estimated.Exports}{Estimated Exports, Total Ind RW}
\item{Adj.Imports}{Adjusted Imports, Total Ind RW}
\item{Adj.Exports}{Adjust Exports, Total Ind RW}
\item{Estimated.Imports(tons)}{Estimated Imports in tons, Total Ind RW}
\item{Estimated.Exports(tons)}{Estimated Exports in tons, Total Ind RW}
\item{Adj.Imports(tons)}{Adjusted Imports in tons, Total Ind RW}
\item{Adj.Exports(tons)}{Adjust Exports in tons, Total Ind RW}
\item{ApparentConsumption.Total}{Apparent Conusmption, Total Ind RW.}
\item{SawLogs.Dom.Prod}{Domestic Production of Saw Logs}
\item{SawLogs.NetImports}{Net Imports of Saw Logs.
Net imports of lumber converted to cubic feet roundwood.  Small quantities of imported saw logs (roundwood form) are included under domestic production.}
\item{SawLogs.ApparentConsumption}{Apparent Consumption of Saw Logs
Includes changes in stocks beginning in 1935.}
\item{VeneerLogs.Dom.Prod}{Domestic Production of Veneer Logs}
\item{VeneerLogs.NetImports}{Net Imports of Veneer Logs
Net imports of veneer logs represent the equivalent net imports of veneer and plywood converted to board feet log scale, and then to cubic feet round-wood.
The small volume of veneer logs imported (roundwood form) is included under domestic production.}
\item{VeneerLogs.ApparentConsumption}{Apparent Consumption of Veneer Logs.}
\item{PulpWood.Dom.Prod}{Domestic Production of Pulpwood}
\item{PulpWood.NetImports}{Net Imports of Pulpwood}
\item{PulpWood.ApparentConsumption}{Apparent Consumption of Pulpwood}
\item{Other.ApparentConsumption}{Apparent Consumption of Other Products}
\item{FuelWood.ApparentConsumption}{Apparent Consumption of Fuelwood}
}}
\source{
US Department of Commerce and Agriculture and USFS Estimations
}
\usage{
hair1963
}
\description{
Dataset containing estimated statistics for timber products. All vectors are numeric.
Most data is for products in the Industrial Roundwood Class. Each column except the first two and the last one
are in this category.
}
\details{
Columns may not add to total because of rounding.
}
\keyword{hair1963}

