% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{hair1963t20}
\alias{hair1963t20}
\title{hair1963t20}
\format{A data frame with 36 observations on 8 variables}
\source{
US Department of Commerce, Bureau of the Census
}
\usage{
hair1963t20
}
\description{
Imports and exports of plywood, by softwood and hardwood in the US between 1935 and 1962 (estimates of 1927-34)
}
\keyword{hair1963t20}

