% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{halfLives}
\alias{halfLives}
\title{Half lives for each end use.}
\format{A data frame with 151 observations of 13 variables:
\describe{
\item{House.SingFam}{Half lives for Single Family Homes}
\item{House.Multifam}{Half lives for Multi Family Homes}
\item{House.MobHom}{Half lives for Mobile Homes}
\item{Res.Upkeep}{Half lives for Total Residential Upkeep and Improvement}
\item{New.Nonres.AllRR}{Half lives for All Railroads}
\item{New.Nonres.Rties}{Half lives for Railroad Ties}
\item{New.Nonres.Rcar.Repair}{Half lives for Railcar Repair}
\item{Manu.HouseFurniture}{Half lives for Household Furniture}
\item{Manu.CommFurniture}{Half lives for Commercial Furniture}
\item{Manu.OtherProducts}{Half lives for Manufacturing and Other Products}
\item{Shipping.Tot}{Half lives for Shipping}
\item{Other.Uses.Tot}{Half lives for Other Uses}
\item{Other.Industrial.Tot}{Half lives for Uses For Other Industrial Products}
}}
\source{
USFS data
}
\usage{
halfLives
}
\description{
Half lives for each of the 13 end uses.
}
\keyword{datasets}

