% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard38}
\alias{howard38}
\title{Production, imports, exports, and consumption of structural panel products, by type, 1980-1999}
\format{A data frame with 41 observations and 21 variables:
\describe{
\item{Prod.Tot}{Total Production}
\item{Prod.SWPlywood}{Softwood Plywood Production}
\item{Prod.OSP}{Other Structural Panels Production}
\item{Imports.Tot}{Total Imports}
\item{Imports.SWPlywood}{Softwood Plywood Imports}
\item{Imports.OSP}{Other Structual Panels Imports}
\item{Exports.Tot}{Total Exports}
\item{Exports.SWPlywood}{Softwood Plywood Exports}
\item{Exports.OSP}{Other Structural Panels Exports}
\item{Consump.Tot}{Total Consumption}
\item{Consump.SWPlywood}{Softwood Plywood Consumption}
\item{Consump.OSP}{Other Structural Panels Consumption}
}}
\source{
American Plywood Association, US Dept. of Agriculture
Foreign Agricultural Service
}
\usage{
howard38
}
\description{
Datasets containing statistics for various structural panel products.
All units are in Million Square ft.
}
\keyword{howard38}

