% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{howard49}
\alias{howard49}
\title{Wood pulp production, imports, exports, and consumption, 1965-1999}
\format{A data frame with 56 observations of 15 variables}
\source{
U.S. Department of Commerce, Bureau of the Census (52,56,57); American Forest
and Paper Association (2,4); United Nations, Food and Agriculture Organization (29).
1993-1997 revised to include dissolving and special alpha pulps, which is consistent with earlier years.
American Paper Institute; American Forest & Paper Association: Statistics of Paper, Paperboard & Wood Pulp;
 1977, 1982, 1990, 1995, 1998, 1999 annual Statistics reports (4).
 1965-1967, 1970, 1973, 1977-1978, 1986-1988 & 1997 revised.
 1966, 1968, 1972, 1975, 1985-1987 & 1996-1997 revised.
}
\usage{
howard49
}
\description{
Dataset containing woodpulp statistics.
All units are in thousand pounds unless otherwise specified.
Consumption = Production + Imports - Exports
}
\details{
\itemize{
\item \strong{Production} Wood pulp Production
\item \strong{Imports} Wood pulp Imports
\item \strong{Imports.PercConsump} Imports as Percent of Consumption
\item \strong{Exports} Wood pulp Exports
\item \strong{Exports.PercProd} Exports as Percent of Production
\item \strong{Consump.Tot} Total Consumption
\item \strong{Consump.PerCap(pounds)} Per Capita Consumption in pounds
}
}
\keyword{datasets}

