% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard53}
\alias{howard53}
\title{Particleboard and medium-density fiberboard production, imports,}
\format{A data frame with 56 observations of 7 variables:
\describe{
\item{Prod.Tot}{Total Particleboard and Medium-Density Fiberboard Production}
\item{Prod.ParticleBoard}{Particleboard Production}
\item{Prod.MediumDensityFiberboard}{Medium-Density Fiberboard Production}
\item{Imports}{Particleboard and Medium-Density Fiberboard Imports}
\item{Exports}{Particleboard and Medium-Density Fiberboard Exports}
\item{Consump.Tot}{Total Particleboard and Medium-Density Fiberboard Consumption}
\item{Consump.PerCap(Sq ft)}{Per Capita Consumption of particleboard and medium-density fiberboard, in Square ft.}
}}
\source{
National Particleboard Association (20); U.S. International Trade Commission (69);
U.S. Department of Agriculture (32).  Data may not add to totals because of rounding.
}
\usage{
howard53
}
\description{
Dataset containing various statistics for particleboard and medium-density fiberboard.
All units are in Million Square ft unless otherwise specified.
}
\keyword{datasets}

