% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard55}
\alias{howard55}
\title{Insulating board production, imports, exports, and consumption, 1965-1999.}
\format{A data frame with 56 observations of  5 variables. (all units in thousand tons except when specified otherwise):
\itemize{
\item \code{Production} (Production)
\item \code{Imports}(Imports)
\item \code{Exports}(Exports)
\item \code{Total}(Total)
\item \code{PerCapita}(Per Capita units: Pounds)
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard55
}
\description{
Insulating board production, imports, exports, and consumption, 1965-1999.
}
\keyword{howard55}

