% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard56}
\alias{howard56}
\title{Hardboard production, imports, exports and consumption, 1965-1999.}
\format{A data frame with 56 observations on 5 variables:
\describe{
\item{Years}{Years}
\item{Hardboard.Production}{Hardboard production}
\item{Hardboard.Imports}{Hardboard imports}
\item{Hardboard.Exports}{Hardboard exports}
\item{Hardboard.Total.Consumption}{Total Hardboard Consumption}
\item{Hardboard.PerCapita.Consumption}{Per Capita Hardboard Consumption in Square feet}
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard56
}
\description{
All units are in million square feet except when specified otherwise.
}
\keyword{datasets}

