% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocSami.R
\docType{data}
\name{howard6a}
\alias{howard6a}
\title{Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.}
\format{A data frame with 56 observations on 34 variables. (all units are in million cubic feet except when specified otherwise)
\itemize{
\item \code{AllProduction.Prod} (All products production)
\item \code{AllProduct.Consump}(All products consumption)
\item \code{Ind.RW.Tot.Prod}(Industrial roundwood uses total production)
\item \code{Ind.RW.Tot.Imports}(Industrial roundwood uses total imports)
\item \code{Ind.RW.Tot.Exports}(Industrial roundwood uses total exports)
\item \code{Ind.RW.Tot.Consump}(Industrial roundwood uses total consumption)
\item \code{Ind.RW.Lum.Prod}(Industrial roundwood uses lumber production)
\item \code{Ind.RW.Lum.Imports}(Industrial roundwood uses lumber imports)
\item \code{Ind.RW.Lum.Exports}(Industrial roundwood uses lumber exports)
\item \code{Ind.RW.Lum.Consump}(Industrial roundwood uses lumber consumption)
\item \code{Ind.RW.PlyandVen.Prod}(Industrial roundwood uses plywood and veneer production)
\item \code{Ind.RW.PlyamdVen.Imports}(Industrial roundwood uses plywood and veneer imports)
\item \code{Ind.RW.PlyandVen.Exports}(Industrial roundwood uses plywood and veneer exports)
\item \code{Ind.RW.PlyandVen.Consump}(Industrial roundwood uses plywood and veneer consumption)
\item \code{IndRW.Pulp.Prod}(Industrial roundwood uses pulp-based products production)
\item \code{Ind.RW.Pulp.Imports}(Industrial roundwood uses pulp-based products imports)
\item \code{Ind.RW.Pulp.Exports}(Industrial roundwood uses pulp-based products exports)
\item \code{Ind.RW.Pulp.Consump}(Industrial roundwood uses pulp-based products consumption)
\item \code{Ind.RW.OtherIndustrial.ProdAndConsump}(Other industrial product production and consumption)
\item \code{Ind.RW.Logs.Imports}(Industrial roundwood uses log imports)
\item \code{Ind.RW.Logs.Exports}(Industrial roundwood uses log exports)
\item \code{Ind.RW.Pulp.Imports}(Industrial roundwood uses pulpwood chip imports)
\item \code{Ind.RW.Pulp.Exports}(Industrial roundwood uses pulpwood chip exports)
\item \code{FuelWood.ProdAndConsumption}(Fuelwood production and consumption)
\item \code{UnNamed1}
\item \code{UnNamed2}
\item \code{UnNamed3}
\item \code{UnNamed4}
\item \code{UnNamed5}
\item \code{UnNamed6}
\item \code{UnNamed7}
\item \code{UnNamed8}
\item \code{UnNamed9}
\item \code{UnNamed10}
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard6a
}
\description{
Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.
}
\keyword{howard6a}

