% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard7}
\alias{howard7}
\title{Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.}
\format{A data frame with 56 observations on 31 variables:
\describe{
\item{AllProduction.Prod}{All products production}
\item{AllProduct.Consump}{All products consumption}
\item{Indu.RW.Tot.Prod}{Industrial roundwood uses total production}
\item{Indu.RW.Tot.Imports}{Industrial roundwood uses total imports}
\item{Indu.RW.Tot.Exports}{Industrial roundwood uses total exports}
\item{Indu.RW.Tot.Consump}{Industrial roundwood uses total consumption}
\item{Indu.RW.Lum.Prod}{Industrial roundwood uses lumber production}
\item{Indu.RW.Lum.Imports}{Industrial roundwood uses lumber imports}
\item{Indu.RW.Lum.Exports}{Industrial roundwood uses lumber exports}
\item{Indu.RW.Lum.Consump}{Industrial roundwood uses lumber consumption}
\item{Indu.RW.PlyandVen.Prod}{Industrial roundwood uses plywood and veneer production}
\item{Indu.RW.PlyandVen.Imports}{Industrial roundwood uses plywood and veneer imports}
\item{Indu.RW.PlyandVen.Exports}{Industrial roundwood uses plywood and veneer exports}
\item{Indu.RW.PlyandVen.Consump}{Industrial roundwood uses plywood and veneer consumption}
\item{Indu.RW.Pulp.Prod}{Industrial roundwood uses pulp-based products production}
\item{Indu.RW.Pulp.Imports}{Industrial roundwood uses pulp-based products imports}
\item{Indu.RW.Pulp.Exports}{Industrial roundwood uses pulp-based products exports}
\item{Indu.RW.Pulp.Consump}{Industrial roundwood uses pulp-based products consumption}
\item{Indu.RW.OtherIndustrial.ProdAndConsump}{Other industrial product production and consumption}
\item{Indu.RW.Logs.Imports}{Industrial roundwood uses log imports}
\item{Indu.RW.Logs.Exports}{Industrial roundwood uses log exports}
\item{Indu.RW.Pulp.Imports}{Industrial roundwood uses pulpwood chip imports}
\item{Indu.RW.Pulp.Exports}{Industrial roundwood uses pulpwood chip exports}
\item{FuelWood.ProdAndConsumption}{Fuelwood production and consumption}
\item{UnNamed1}{}
\item{UnNamed2}{}
\item{UnNamed3}{}
\item{UnNamed4}{}
\item{UnNamed5}{}
\item{UnNamed6}{}
\item{UnNamed7}{}
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard7
}
\description{
All units are in million cubic feet except when specified otherwise.
}
\keyword{datasets}

