% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{IncePaper}
\alias{IncePaper}
\title{U.S. Paper & Paperboard, Production, Imports, Exports, Apparent Consumption, 1900 - 2050
Data set containing statistics for paper and paperboard in the U.S.}
\format{An object of class \code{data.frame} with 152 rows and 10 columns.}
\usage{
IncePaper
}
\description{
U.S. Paper & Paperboard, Production, Imports, Exports, Apparent Consumption, 1900 - 2050
Data set containing statistics for paper and paperboard in the U.S.
}
\keyword{datasets}

