% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lossIU}
\alias{lossIU}
\title{Fraction of wood lost when placed in use from various end uses.}
\format{A data frame with 151 observations of 16 variables:
\describe{
\item{Years}{Years}
\item{House.SingFam}{Fraction of wood lost in single family homes}
\item{House.Multifam}{Fraction of wood lost in multi family homes}
\item{House.MobHom}{Fraction of wood lost in mobile homes}
\item{Res.Upkeep}{Fraction of wood lost in total residential upkeep and improvement}
\item{New.Nonres.AllRR}{Fraction of wood lost in all railroads}
\item{New.Nonres.Rties}{Fraction of wood lost in railroad ties}
\item{New.Nonres.Rcar.Repair}{Fraction of wood lost in railcar repair}
\item{Manu.HouseFurniture}{Fraction of wood lost in household furniture}
\item{Manu.CommFurniture}{Fraction of wood lost in commercial furniture}
\item{Manu.OtherProducts}{Fraction of wood lost in manufacturing and other products}
\item{Shipping.Tot}{Fraction of wood lost in shipping}
\item{Other.Uses.Tot}{Fraction of wood lost in other uses}
\item{Other.Industrial.Tot}{Fraction of wood lost in other industrial products}
}}
\source{
Check spreadsheet - values entered in
}
\usage{
lossIU
}
\description{
All units are percentages and all vectors are numeric.
}
\keyword{datasets}

