% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{lumberpre1900}
\alias{lumberpre1900}
\title{Lumber wood  in single and multifamily houses build 1800 to 1900}
\format{A vector with 151 observations of 1 variable.

\itemize{
\item \strong{Carbon} Lumber wood  in single and multifamily houses
}}
\source{
'ftp://ftp2.fs.fed.us/incoming/fpl/Skog/[Woodcarb II special lumber July 5 2007 Base Case.xls]SW Calc'!$V3878
}
\usage{
lumberpre1900
}
\description{
All vectors are numeric. All units are ?.
}
\keyword{lumberpre1900}

