% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{paper_carbon_stockchange}
\alias{paper_carbon_stockchange}
\title{Calculates changes in carbon stock from paper.
Units are in Tg Carbon.
For production approach, values correspond to `Calculation$CC` in the spreadsheet.
For stock change approach, values correspond to `Calculation$L` in the spreadsheet.}
\usage{
paper_carbon_stockchange(years, approach = c("Production", "Stock Change"))
}
\arguments{
\item{years}{years to return carbon totals for}

\item{approach}{The approach used to calculate carbon contribution.}
}
\value{
A vector of paper carbon stock changes for `years`
}
\description{
Calculates changes in carbon stock from paper.
Units are in Tg Carbon.
For production approach, values correspond to `Calculation$CC` in the spreadsheet.
For stock change approach, values correspond to `Calculation$L` in the spreadsheet.
}
\examples{
\dontrun{
paper_carbon_stockchange(1990:2000)
paper_carbon_stockchange(1950:1975, approach = "Stock Change")
}
}

