% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{paperToLandFills}
\alias{paperToLandFills}
\title{Percent of paper that goes to landfills.}
\format{A data frame with 151 observations of 1 variable.

\itemize{
\strong{PaperToLandfills} Percent of paper that goes to landfills
}}
\source{
Linked site, check spreadsheet
}
\usage{
paperToLandFills
}
\description{
Units are percentages and the vector is numeric.
}
\keyword{paperToLandFills}

