% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{plf_percent}
\alias{plf_percent}
\title{Percent of paper that goes to landfills.}
\format{A data frame with 151 observations of 1 variable:
\describe{
\item{Years}{Years}
\item{plf_percent}{Percent of paper that goes to landfills}
}}
\source{
Linked site, check spreadsheet
}
\usage{
plf_percent
}
\description{
Units are percentages and the vector is numeric.
}
\keyword{datasets}
