% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich29}
\alias{ulrich29}
\title{Lumber production, imports, exports, and consumption, by softwoods and hardwoods, 1950-87}
\format{A data frame with 38 observations of 15 variables:
\describe{
\item{Years}{Years}
\item{Prod.Tot}{Total Lumber Production}
\item{Prod.SW}{Softwood Lumber Production}
\item{Prod.HW}{Hardwood Lumber Prodcution}
\item{Imports.Tot}{Total Lumber Imports}
\item{Imports.SW}{Softwood Lumber Imports}
\item{Imports.HW}{Hardwood Lumber Imports}
\item{Exports.Tot}{Total Lumber Exports}
\item{Exports.SW}{Softwood Lumber Exports}
\item{Exports.HW}{Hardwood Lumber Exports}
\item{Consump.Tot}{Total Lumber Consumption}
\item{Consump.SW}{Softwood Lumber Consumption}
\item{Consump.HW}{Hardwood Lumber Consumption}
\item{PerCapitaConsump.Tot}{Total Lumber Consumption Per Capita, in Board ft.}
\item{PerCapitaConsump.SW}{Softwood Lumber Consumption Per Capita, in Board ft.}
\item{PerCapitaConsump.HW}{Hardwood Lumber Consumption Per Capita, in Board ft.}
}}
\source{
U.S. Department of Commerce. Bureau of the Census (43.51.52)
U.S. Department of Agriculture, Forest Service; National Forest Products Association (16).
Data for selected years 1899.1949 in "The demand and price situation for forest products", 1964, table 7 (31)
}
\usage{
ulrich29
}
\description{
Dataset containing various statistics for lumber products.
All units are in Billion Board Ft unless otherwise specified.
Year 1987 includes small volumes of mixed species (not classified as softwoods or hardwoods).
Data may not add to totals because of rounding. Estimates of softwood lumber production for certain years in the
late 1970's and early 1980's by the Bureau of the Census and some major softwood industry associations currently vary by as much as
5 percent. Cooperative studies are underway to resolve these differences. Revisions made by Census as a result of these studies
will be reported in future issues of U.S. timber production, trade, consumption, and price statistics.
}
\keyword{datasets}

