% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich5}
\alias{ulrich5}
\title{Production, imports, exports, and consumption of softwood timber products, by major product, 1950-87}
\format{A data frame with 38 observations of 24 variables:
\describe{
\item{Years}{Years}
\item{AllProducts.Prod}{Production for all timber products}
\item{AllProducts.Consump}{Consumption for all timber products}
\item{Tot.Prod}{Total Production for major timber products}
\item{Tot.Imports}{Total Imports for major timber products}
\item{Tot.Exports}{Total Exports for major timber products}
\item{Tot.Consump}{Total Consumption for major timber products}
\item{Lumber.Prod}{Lumber Products Prodcution}
\item{Lumber.Imports}{Lumber Products Imports}
\item{Lumber.Exports}{Lumber Products Exports}
\item{Lumber.Consump}{Lumber Products Consumption}
\item{PlywoodVeneer.Prod}{Plywood and Veneer Products Production}
\item{PlywoodVeneer.Imports}{Plywood and Veneer Products Imports}
\item{PlywoodVeneer.Exports}{Plywood and Veneer Products Exports}
\item{PlywoodVeneer.Consump}{Plywood and Veneer Products Consumption}
\item{PulpProducts.Prod}{Pulp Products Production}
\item{PulpProducts.Imports}{Pulp Products Imports}
\item{PulpProducts.Exports}{Pulp Products Exports}
\item{PulpProducts.Consump}{Pulp Products Consumption}
\item{OtherIndProducts.ProdandConsump}{Production and Consumption for other industrial products}
\item{Logs.Imports}{Log Imports}
\item{Logs.Exports}{Log Exports}
\item{PulpwoodChip.Exports}{Pulpwood-chip Exports}
\item{Fuelwood.ProdandConsump}{Production and Consumption for fuelwood products}
\item{LogChipExports.PercofProduction}{Log and Chip exports as a percent of production}
}
Based on data published by the U.S. Departments of Commerce and Agriculture.
Data for 1900-49 in The demand and price situation for forest products 1964, table 2 (31),}
\usage{
ulrich5
}
\description{
Dataset containing statistics for softwood timber products.
All units are in Million cubic ft(roundwood equivalent). All vectors
are numeric.
}
\keyword{datasets}

