% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich52}
\alias{ulrich52}
\title{Particleboard and medium density fiberboard production, imports, exports, and consumption, 1950-87}
\format{A data frame with 38 observations of  7 variables:
\describe{
\item{Years}{Years}
\item{Prod.tot}{Production Total}
\item{Prod.Part.Board}{Production of Partical Board}
\item{Prod.Med.Fiberboard}{Medium Density Fiberboard}
\item{Imports}{Particleboard and medium density fiberboard imports}
\item{Exports}{Particleboard and medium density fiberboard exports}
\item{Consump.Tot}{Consumption Total}
\item{Consump.PerCapita}{Consumption Per Capita, square feet}
}}
\source{
U.S. Department of Commerce, Bureau of the Census
}
\usage{
ulrich52
}
\description{
All units in millions square feet except when specified otherwise.
}
\keyword{datasets}

