% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich53}
\alias{ulrich53}
\title{Insulating board production, imports, exports, and consumption, 1950-87}
\format{A data frame with 63 observations of  6 variables. (all units in millions square feet except when specified otherwise):
\itemize{
\item \code{InsulatingBoard.Prod} (Insulating Board Production)
\item \code{InsulatingBoard.Import}(Insulating Board Imports)
\item \code{InsulatingBoard.Exports}(Inulating Board Exports)
\item \code{InsulatingBoard.Consump.Tot}(Insulating Board consumption total)
\item \code{InsulatingBoard.Consump.PerCapita}(Insulating Board consumption per capita (square feet))
}}
\source{
U.S. Department of Commerce, Bureau of the Census
}
\usage{
ulrich53
}
\description{
Insulating board production, imports, exports, and consumption, 1950-87
}
\keyword{ulrich53}

