% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{ulrich54}
\alias{ulrich54}
\title{Production, imports, exports and consumption of insulating board, 1950-1987
Also includes historical production and exports data from 1925-1950.}
\format{A data frame with 72 observations of 5 variables}
\source{
U.S. Department of Commerce, Bureau of the Census (48,51, 52);
American Paper Institute (3, 4); American Hardboard Association.
Data from Ince,2000, Industrial Wood Productivity in the United States, 1900-1998.
}
\usage{
ulrich54
}
\description{
Dataset contains statistics for insulating board.
All units are in million square ft unless otherwise specified. All vectors are numeric.
}
\details{
\itemize{
\item \strong{HardBoard.Prod} HardBoard Production
\item \strong{HardBoard.Import} HardBoard Imports
\item \strong{HardBoard.Exports} HardBoard Exports
\item \strong{HardBoard.Consump.Tot} Total HardBoard Consumption
\item \strong{HardBoard.Consump.PerCapita} HardBoard Consumption Per Capita
}
}
\keyword{datasets}

