% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich6}
\alias{ulrich6}
\title{Production, imports, exports, and consumption of hardwood timber products, by major product, 1950-87}
\format{A data frame with 48 observations of 22 variables:
\describe{
\item{AllProduction.Prod}{Total production of all products}
\item{AllProduct.Consump}{Total consumption of all products}
\item{Indu.RW.Tot.Prod}{Total production of all industrial roundwood products}
\item{Indu.RW.Tot.Imports}{Total imports of all industrial roundwood products}
\item{Indu.RW.Tot.Exports}{Total exports of all industrial roundwood products}
\item{Indu.RW.Tot.Consump}{Total consumption of all industrial roundwood products}
\item{Indu.RW.Lum.Prod}{Production of lumber products}
\item{Indu.RW.Lum.Imports}{Imports of lumber products}
\item{Indu.RW.Lum.Exports}{Exports of lumber products}
\item{Indu.RW.Lum.Consump}{Consumption of lumber products}
\item{Indu.RW.PlyandVen.Prod}{Production of plywood and veneer products}
\item{Indu.RW.PlyandVen.Imports}{Imports of plywood and veneer products}
\item{Indu.RW.PlyandVen.Exports}{Exports of plywood and veneer products}
\item{Indu.RW.PlyandVen.Consump}{Consumption of plywood and veneer products}
\item{Indu.RW.Pulp.Prod}{Production of pulp products}
\item{Indu.RW.Pulp.Imports}{Imports of pulp products}
\item{Indu.RW.Pulp.Exports}{Exports of pulp products}
\item{Indu.RW.Pulp.Consump}{Consumption of pulp products}
\item{Indu.RW.OtherIndustrial.ProdAndConsump}{Production and consumption of Other industrial roundwood products}
\item{Indu.RW.Logs.Imports}{Imports of logs}
\item{Indu.RW.Logs.Exports}{Exports of logs}
\item{FuelWood.ProdAndConsumption}{Production and Consumption of Fuelwood}
\item{UnNamed1}{Unnamed}
}}
\source{
Based on data published by the U.S. Departments of Commerce and Agriculture.
}
\usage{
ulrich6
}
\description{
Dataset containing statistics for hardwood timber products.
All units are in million cubic ft.
}
\keyword{datasets}

