% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{usaFiberPulp}
\alias{usaFiberPulp}
\title{Recovered Fibre Pulp in the United States from 1998 to 2013}
\format{A data frame with 16 observations of 3 variables

\itemize{
\strong{Prod.Quantity} Production Quantity of Recovered Fibre Pulp
\strong{Imports.Quantity} Imports Quantity of Recovered Fibre Pulp
\strong{Exports.Quantity} Exports Quantity of Recovered Fibre Pulp
}}
\source{
from AF&PA Paper, paperboard & wood pulp, A Monthly statistical summary
}
\usage{
usaFiberPulp
}
\description{
All units are in metric tons and all vectors are numeric.
}
\keyword{usaFiberPulp}

