% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcarbon.R
\name{Var1_C_SWP_STOCKCHANGE}
\alias{Var1_C_SWP_STOCKCHANGE}
\title{Title}
\usage{
Var1_C_SWP_STOCKCHANGE(years, decaydistribution = c("Exponential", "Gamma"))
}
\arguments{
\item{decaydistribution}{}
}
\value{

}
\description{
Title
}

