% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papercarbon.R
\name{Var2_C_PAPER_STOCKCHANGE}
\alias{Var2_C_PAPER_STOCKCHANGE}
\title{Title}
\usage{
Var2_C_PAPER_STOCKCHANGE(y)
}
\arguments{
\item{y}{year for stock change}
}
\value{
Stock change in carbon from paper
}
\description{
Carbon stock change for paper production year y in Tg C/yr
}

