% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcarbon.R
\name{Var2_C_SWP_STOCKCHANGE}
\alias{Var2_C_SWP_STOCKCHANGE}
\title{Carbon Stock Change of Solidwood Products}
\usage{
Var2_C_SWP_STOCKCHANGE(years, decaydistribution = c("Exponential", "Gamma"))
}
\arguments{
\item{year}{the year to calculate stock change for}

\item{distribution}{the decay distribution to be used in the calculation}
}
\value{

}
\description{
Carbon Stock Change of Solidwood Products
}

