% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{Var2b_C_STOCKCHANGE_TOTAL_2B}
\alias{Var2b_C_STOCKCHANGE_TOTAL_2B}
\title{VAR #2b
Annual Change in stock of HWP in SWDS produced from domestic harvest
Output in Gg C/yr}
\usage{
Var2b_C_STOCKCHANGE_TOTAL_2B(y)
}
\description{
VAR #2b
Annual Change in stock of HWP in SWDS produced from domestic harvest
Output in Gg C/yr
}

