% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvest.R, R/var5.R
\name{var5}
\alias{var5}
\title{Annual Domestic Harvest}
\usage{
var5(years, onlyvar = FALSE)

var5(years, onlyvar = FALSE)
}
\arguments{
\item{years}{years to calculate statistics for}

\item{onlyvar}{If true, only the variable 5 will be returned.
If false, other calculations needed to produce variable 5 will be returned}

\item{onlyvar}{}
}
\value{
Depends on onlyvar, either data frame of intermediate calculations or a vector
of values for Annual Domestic Harvest
}
\description{
Various statistics and intermediate calculations to solve for
Variable 5, Annual Domestic Harvest (H)

Variable 5
}

