% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wd_percent}
\alias{wd_percent}
\title{Percent of wood to dumps}
\format{A data frome with 151 observations of 1 variable:

\describe{
\item{Years}{Years}
\item{wd_percent}{Percentage of wood that goes to dumps}
}}
\source{
Linked spreadsheet (check this)
}
\usage{
wd_percent
}
\description{
Gives percentage of all wood that goes to dumps.
}
\keyword{datasets}

