% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wlf_percent}
\alias{wlf_percent}
\title{Percentage of wood to landfills.
Corresponds to colum U20 in Param + Results Sheet.
Gives percentage of all wood that goes to landfills.}
\format{A data frame with 151 observations of 1 variable:
\describe{
\item{Years}{Years}
\item{wlf_percent}{Percentage of wood going to landfills}
}}
\source{
Linked spreadsheet (check this)
}
\usage{
wlf_percent
}
\description{
Percentage of wood to landfills.
Corresponds to colum U20 in Param + Results Sheet.
Gives percentage of all wood that goes to landfills.
}
\keyword{datasets}

