% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{woodToDumps}
\alias{woodToDumps}
\title{Percent of wood that goes to dumps}
\format{A data frome with 151 observations of 1 variable

\itemize{
\strong{WoodToDumps} Percentage of wood that goes to dumps
}}
\source{
Linked spreadsheet (check this)
}
\usage{
woodToDumps
}
\description{
All units are percentages and the vector is numeric
}
\keyword{woodToDumps}

