% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmosphericflow_harvest.R
\name{calcP_EX}
\alias{calcP_EX}
\title{Statistics for variable 4}
\usage{
calcP_EX(years = 1990:2020, var = FALSE, woodToCarbon = 4.535925e-07,
  paperToCarbon = 3.9008955e-07)
}
\arguments{
\item{years}{years to calculate}

\item{var}{If true, return only variable 4. If false, return intermediate statistics needed for variable 4}
}
\value{
if var = FALSE, necessary calculations for Variable 4 for selected years
        if var = TRUE, returns values for Variable 4 for selected years
}
\description{
Statistics for variable 4
}

