% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmosphericflow.R
\name{calcP_IM}
\alias{calcP_IM}
\title{Statistics needed for variable 3
Calculates intermediate calculations for variable 3:
Annual Imports of Wood, and paper products + wood fuel,
pulp, recovered paper, roundwood/chips.}
\usage{
calcP_IM(years = 1990:2020, var = FALSE)
}
\arguments{
\item{years}{years to calculate for}

\item{var}{if true, returns variable 3. if false, returns statistics necessary to calculate variable 3}
}
\value{
if var = FALSE, necessary intermediate calculations for Variable 3
        if var = TRUE, returns values for Variable 3 for selected years
}
\description{
Statistics needed for variable 3
Calculates intermediate calculations for variable 3:
Annual Imports of Wood, and paper products + wood fuel,
pulp, recovered paper, roundwood/chips.
}

