% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericFlow.R
\name{calcP_IM}
\alias{calcP_IM}
\title{Statistics for variable 3}
\usage{
calcP_IM(years = 1990:2020, var = FALSE)
}
\arguments{
\item{years}{years to calculate for}

\item{var}{if true, returns variable 3. if false, returns statistics necessary to calculate variable 3}
}
\value{
if var = FALSE, necessary intermediate calculations for Variable 3
        if var = TRUE, returns values for Variable 3 for selected years
}
\description{
Statistics for variable 3
}
\examples{
calcP_IM(years = 1950:1980)
calcP_IM(var = TRUE)
calcP_IM(years = 2012:2015, var = TRUE)
}

