% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swp-carbon.R
\name{calcplacediu}
\alias{calcplacediu}
\title{Title}
\usage{
calcplacediu(Years = 1900:2020, approach = c("Stock Change", "Production"),
  total = TRUE)
}
\value{
a data frame of carbon placed in use for various end uses
}
\description{
Title
}
\examples{
calcplacediu()
}

