% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swp.R
\name{calcplacediu}
\alias{calcplacediu}
\title{Total carbon placed in use in od tons of wood fiber}
\usage{
calcplacediu(Years = 1900:2020, approach = c("Production", "Stock Change"),
  total = TRUE, fsp = fracstrpanels, fnsp = fracnonstrpanels,
  fsawn = fracsawnwood, swpdata)
}
\arguments{
\item{Years}{years to calculate carbon placed in use for}

\item{approach}{The approach type to use. Fractions dont change but inflow does}

\item{total}{Whether to return total placed in use or seperate it by end uses.}

\item{fsp}{Fraction of structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fnsp}{Fraction of non-structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fsawn}{Fraction of sawnwood products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{swpdata}{Optional argument to provide swp production and trade time series
for purposes of error analysis}
}
\value{
a data frame of carbon placed in use for various end uses
}
\description{
Corresponds to B3147 column in SW Calc P and SW Calc spreadsheet for Production and Stock Change
approach respectively.
}

