% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWPstatistics.R
\name{calcplacediu}
\alias{calcplacediu}
\title{Total carbon placed in use in od tons of wood fiber}
\usage{
calcplacediu(Years = 1900:2020, approach = c("Production", "Stock Change"),
  total = TRUE)
}
\arguments{
\item{Years}{years to calculate carbon placed in use for}

\item{approach}{The approach type to use. Fractions dont change but inflow does}

\item{total}{Whether to return total placed in use or seperate it by end uses.}
}
\value{
a data frame of carbon placed in use for various end uses
}
\description{
Corresponds to B3147 column in SW Calc P and SW Calc spreadsheet for Production and Stock Change
approach respectively.
}

