% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpsvar2b.R
\name{calculatedumpcarbonproduction}
\alias{calculatedumpcarbonproduction}
\title{Calculate carbon in SWDS for the Stock Change Method}
\usage{
calculatedumpcarbonproduction(Years = yrs, totalcarbonstockchange = FALSE)
}
\value{
Either a data frame of values from Calculation spreadsheet or total carbon output
}
\description{
Column names with only letters come from Calc sheet, otherwise sheet is specified (need to change names)
}
\details{
Calculates carbon in SWDS with production approach
}

