% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{calculatedumpcarbonproduction}
\alias{calculatedumpcarbonproduction}
\title{Calculate carbon in SWDS for the Production Method}
\usage{
calculatedumpcarbonproduction(Years, totalcarbonstockchange = FALSE)
}
\arguments{
\item{Years}{years to calculate}

\item{totalcarbonstockchange}{If true, returns carbon stock change, or returns data
frame of intermediate calculations}
}
\value{
Either a data frame of values from Calculation spreadsheet or total carbon output
}
\description{
Column names with only letters come from Calc sheet, otherwise sheet is specified (need to change names)
}

