% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpsvar1b.R
\name{calculatedumpcarbonstockchange}
\alias{calculatedumpcarbonstockchange}
\title{Calculate carbon in SWDS for the Stock Change Method}
\usage{
calculatedumpcarbonstockchange(Years = yrs, totalcarbonstockchange = FALSE)
}
\arguments{
\item{totalcarbonstockchange}{Whether to return data frame of intermedate calculations or
just totals}
}
\value{
Either a data frame of calculations if totalcarbonstockchange is false, or
 a numeric vector of stockchange totals if that argument is true
}
\description{
Calculate carbon in SWDS for the Stock Change Method
}
\examples{
calculatedumpcarbonstockchange()
calculatedumpcarbonstockchange(totalcarbonstockchange = TRUE)
}

