% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumps.R
\name{calculatedumpcarbonstockchange}
\alias{calculatedumpcarbonstockchange}
\title{Calculate carbon in SWDS for the Stock Change Method}
\usage{
calculatedumpcarbonstockchange(Years, totalcarbonstockchange = FALSE)
}
\arguments{
\item{Years}{years to calculate}

\item{totalcarbonstockchange}{Whether to return data frame of intermedate calculations or
just totals}
}
\value{
If totalcarbonstockchange is false: a data frame of intermediate
 calculations.  If totalcarbonstockchange is true: a numeric vector of stock change
 for total carbon from dumps.
}
\description{
Calculate carbon in SWDS for the Stock Change Method
}

