% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SWDSstatistics.R
\name{calculatedumpcarbonstockchange}
\alias{calculatedumpcarbonstockchange}
\title{Calculate carbon in SWDS for the Stock Change Method}
\usage{
calculatedumpcarbonstockchange(Years, totalcarbonstockchange = FALSE)
}
\arguments{
\item{Years}{years to calculate}

\item{totalcarbonstockchange}{Whether to return data frame of intermedate calculations or
 just totals}
}
\value{
Either a data frame of calculations if totalcarbonstockchange is false, or
 a numeric vector of stockchange totals if that argument is true
}
\description{
Calculate carbon in SWDS for the Stock Change Method
}
\examples{
calculatedumpcarbonstockchange(Years = 1990:2015)
calculatedumpcarbonstockchange(Years = 1990:2015, totalcarbonstockchange = TRUE)
}

