% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swp-carbon.R
\name{calculateswp}
\alias{calculateswp}
\title{Intermediate Calculations used for variables 1 and 2, calculated for all years}
\usage{
calculateswp()
}
\value{
Data frame of Roundwood and solidwood production, imports and exports
corresponds to columns in `SWPcalcdata` spreadsheet.
}
\description{
Columns in SWPcalcdata that are used to calculate production and stock change approach.
}
\examples{
calculateswp()
}

