% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWPstatistics.R
\name{calculateswpdata}
\alias{calculateswpdata}
\title{Solidwood products statistics}
\usage{
calculateswpdata()
}
\value{
Data frame of Roundwood and solidwood production, imports and exports
corresponds to columns in `SWPcalcdata` spreadsheet.
}
\description{
Various production, imports and export data for different wood product types.
21 columns in SWPcalcdata that are used to calculate production and stock change approach.
}

