% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{calcUSApaper}
\alias{calcUSApaper}
\title{Calculates paper statistics}
\usage{
calcUSApaper()
}
\value{
A data frame with necessary intermediate calculations to calculate carbon from paper.
Corresponds with necessary columns from the `USA` sheet in the WOODCARB spreadsheet (list them?)
}
\description{
Uses anonymous functions to calculate columns from `USA` sheet in WOODCARB spreadsheet.
}

