% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{finalCarbonContribution}
\alias{finalCarbonContribution}
\title{Calculate final carbon contribution to AFOLU.
All units returned are in Thousand metric tons CO2 Equivalent}
\usage{
finalCarbonContribution(Years = 1990:2015, approach = c("Production",
  "Stock Change"), decaytype = c("Exponential", "K=2"), plot = FALSE,
  halflives = halfLives)
}
\arguments{
\item{Years}{years to return carbon totals for}

\item{approach}{The approach used to calculate carbon contribution.}

\item{decaytype}{Type of decay method to use}

\item{plot}{whether to return a plot or not}

\item{halflives}{data frame of half lives to use}
}
\value{
A vector of carbon contributions for all Years. Plot returned is optional
}
\description{
The 'Production' approach corresponds to `06 IPCC Tables`$R9 in the WOODCARB spreadsheet
The 'Stock Change' approach corresponds to `06 IPCC Tables`$P9 in the WOODCARB spreadsheet
}
\details{
The plot argument returns a simple plot of time vs. carbon contribution.
}
\examples{
finalCarbonContribution()
finalCarbonContribution(approach = 'Stock Change')
finalCarbonContribution(approach = 'Production',
                         decaytype ='Exponential')
}

