% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{finalCarbonContribution}
\alias{finalCarbonContribution}
\title{Calculate final carbon contribution to AFOLU.}
\usage{
finalCarbonContribution(Years = 1990:2015, approach = c("Production",
  "Stock Change"), decaytype = c("Exponential", "Gamma"), plot = FALSE,
  halflives = halfLives)
}
\arguments{
\item{Years}{years to return carbon totals for}

\item{approach}{The approach used to calculate carbon contribution.}

\item{decaytype}{Type of decay method to use}
}
\value{
A vector of carbon contributions for all Years. Plot returned is optional
}
\description{
Calculate final carbon contribution to AFOLU.
}
\examples{
finalCarbonContribution()
finalCarbonContribution(approach = "Stock Change")
finalCarbonContribution(approach = "Production",
                         decaytype ="Exponential")
}

