% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{finalVariables}
\alias{finalVariables}
\title{Calculates the 7 HWP Variables.}
\usage{
finalVariables(Years = 1990:2015, decaydistribution = c("Exponential",
  "K=2"), halflives = halfLives, approach = NULL)
}
\arguments{
\item{Years}{years to calculate}

\item{decaydistribution}{decay distribution to use. Defaults to exponential distribution.}

\item{halflives}{data frame of half lives to use. Must have a column associated with each
one of 13 end uses and half life values for 1900 until the maximum year of interest.}

\item{approach}{approach to calculate corresponding final variables for.
If null, all final variables are calculated and returned.}
}
\value{
If an approach is specified then only the variables needed for that
approach will be returned. If approach is null then a data frame with nine
columns (Variables 1 and 2 have two parts each) will be returned. If the decay distribution
is exponential then the data frame returned will correspond to table containing
final variables: `06 IPCC Tables` in the WOODCARB II spreadsheet.
}
\description{
Returns variables needed to calculate overall contribution.
}
\details{
All units are in Gg Carbon.

Var1A = Annual Change in stock of HWP in use form consumption

Var1B = Annual Change in stock of HWP in SWDS from consumption

Var2A = Annual Change in stock of HWP in use produced from domestic harvest

Var2B = Annual Change in stock of HWP in SWDS produced from domestic harvest

Var3 = Annual Imports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips

Var4 = Annual Exports of wood, paper products, wood fuel, pulp, recovered paper, roundwood/chips

Var5 = Annual Domestic Harvest

Var6 = Annual release of carbon to the atmosphere from HWP consumption (from fuelwood & products
in use and products in SWDS)

Var7 = Annual release of carbon to the atmosphere from HWP (including fuelwood) where wood came from
domestic harvest (from products in use and products in SWDS)
}
\examples{
finalVariables(Years = 1950:2000)
finalVariables(halflives = halfLives * 1.25)
}

