% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{hair1958}
\alias{hair1958}
\title{hair1958}
\format{A data frame with 58 observations on 23 variables (units are in Million Board feet except for percentages):
\itemize{
\item \code{Prod.Tot} (Production Total, numeric vector, in Million Board Ft)
\item \code{Prod.SW} (Production of Softwoods, numeric vector, in Million Board Ft)
\item \code{Prod.HW} (Production of Hardwoods, numeric vector, in Million Board Ft)
\item Imports.Tot. Total Imports, numeric vector, in Million Board ft
\item Imports.SW. Imports for softwoods, in Mil. Board Ft.
\item Imports.HW.
\item Imports.Mixed.
\item Imports.Mixed.PercOfTot.
\item Imports.SW.PerOfTot.
\item Imports.HW.PercOfTot.
\item Imports.Estimated.SW.
\item Imports.Estimated.HW.
\item Exports.Tot.
\item Exports.SW.
\item Exports.HW.
\item Exports.Mixed.
\item Exports.Mixed.PercOfTot.
\item Exports.SW.PercOfTot.
\item Exports.HW.PercOfTot.
\item Exports.Estimated.SW.
\item Exports.Estimated HW.
\item NewSupply.
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
hair1958
}
\description{
Production, imports, exports, and new supply of lumber in US between 1799 and 1958
}
\keyword{hair1958}

