% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{hair1963}
\alias{hair1963}
\title{hair1963}
\format{A data frame with 85 observations on 38 variables}
\source{
US Department of Commerce and Agriculture and USFS Estimations
}
\usage{
hair1963
}
\description{
Estimated production, net imports, and apparent consumption of timber products in the US, 1900-1962

\itemize{
\item Dom.Prod.Tot.
\item ApparentConsumption.
\item Dom.Prod.
}
}
\keyword{hair1963}

