% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hair1963t20}
\alias{hair1963t20}
\title{United States imports and exports of plywood, by softwoods and hardwoods, 1935-62 (with estimates of 1927-34)}
\format{A data frame with 36 observations on 8 variables:
\describe{
\item{Years}{Years}
\item{Imports.Tot}{Total Plywood Imports}
\item{SW.Imports}{Softwood Imports}
\item{HW.Imports.Tot}{Total Hardwood Imports}
\item{HW.Imports.Birch}{Birch Imports}
\item{HW.Imports.Other}{Other Hardwoods Imports}
\item{Exports.Tot}{Total Plywood Exports}
\item{SW.Exports}{Softwood Exports}
\item{HW.Exports}{Hardwood Exports}
}}
\source{
US Department of Commerce, Bureau of the Census
}
\usage{
hair1963t20
}
\description{
Dataset containing import and export statistics for plywood products.
All units are in Thousand Sq Ft. All vectors are numeric.
}
\keyword{datasets}

