% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{hair1963t21}
\alias{hair1963t21}
\title{hair1963t21}
\format{A data frame with 36 observations on 6 variables}
\source{
US Department of Commerce, Bureau of the Census
}
\usage{
hair1963t21
}
\description{
Imports and exports of veneer, by kind, in the US 1935-1962 (estimates of 1927-34)
}
\keyword{hair1963t21}

