% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{halfLives}
\alias{halfLives}
\title{Half Lives}
\format{A data frame with 151 observations of 13 variables.
\itemize{
\item \strong{HL.House.SF} Single Family Homes
\item \strong{HL.House.MultF} Multi Family Homes
\item \strong{HL.House.MobHome} Mobile Homes
\item \strong{ResUpKeep.Tot} Total Residential Upkeep and Improvement
\item \strong{NonRes.construc.allRR} All Railroads
\item \strong{NonRes.construc.RR.ties} Railroad Ties
\item \strong{NonRes.construc.Railcar} Railcar Repair
\item \strong{Manuf.House.Furn} Household Furniture
\item \strong{Manuf.Comm.Furn} Commercial Furniture
\item \strong{Manuf.other} Manufacturing - Other Products
\item \strong{ship.tot} Shipping
\item \strong{Other.tot} Other Uses
\item \strong{other.industrial.tot} Uses For Other Industrial Products
}}
\source{
USFS data
}
\usage{
halfLives
}
\description{
Half lives for different end uses.
}
\keyword{halfLives}

