% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{howard28}
\alias{howard28}
\title{Lumber production, imports, exports, and consumption, by softwoods and hardwoods, 1965-1999.}
\format{A data frame with 56 observations of 15 variables}
\source{
None Listed (need to find)
}
\usage{
howard28
}
\description{
Dataset containing statistics for lumber products.
All units are in Billion Board ft. unless otherwise specified. All vectors are numeric.
\itemize{
\item \strong{Prod.Tot} Total Lumber Production
\item \strong{Prod.SW} Softwood Lumber Production
\item \strong{Prod.HW} Hardwood Lumber Production
\item \strong{Imports.Tot} Total Lumber Imports
\item \strong{Imports.SW} Softwood Lumber Imports
\item \strong{Imports.HW} Hardwood Lumber Imports
\item \strong{Consump.Tot} Total Lumber Consumption
\item \strong{Consump.SW} Softwood Lumber Consumption
\item \strong{Consump.HW} Hardwood Lumber Consumption
\item \strong{PerCapConsump.Tot} Total Per Capita Consumption, in Board ft.
\item \strong{PerCapConsump.SW} Softwood Per Capita Consumption, in Board ft.
\item \strong{PerCapConsump.HW} Hardwood Per Capita Consumption, in Board ft.
}
}
\keyword{howard28}

