% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard37}
\alias{howard37}
\title{Plywood production, imports, exports, and consumption, by softwoods and hardwoods, 1965-1999}
\format{A data frame with 56 observations of 15 variables:
\describe{
\item{Years}{Years}
\item{Prod.Tot}{Total Plywood Production}
\item{Prod.SW}{Softwood Plywood Production}
\item{Prod.HW}{Hardwood Plywood Production}
\item{Imports.Tot}{Total Plywood Imports}
\item{Imports.SW}{Softwood Plywood Imports}
\item{Imports.HW}{Hardwood Plywood Imports}
\item{Exports.Tot}{Total Plywood Exports}
\item{Exports.SW}{Softwood Plywood Exports}
\item{Exports.HW}{Hardwood Plywood Exports}
\item{Consump.Tot}{Total Plywood Consumption}
\item{Consump.SW}{Softwood Plywood Consumption}
\item{Consump.HW}{Hardwood Plywood Consumption}
\item{PerCapConsump.Tot}{Total Per Capita Consumption, in Square ft.}
\item{PerCapConsump.SW}{Softwood Per Capita Consumption, in Square ft.}
\item{PerCapConsump.HW}{Hardwood Per Capita Consumption, in Square ft.}
}}
\source{
US Dept. of Commerce, Bureau of Census, American Plywood Association
The Engineered Wood Association, US Dept. of Agriculture, Foreign Agriculture Service
}
\usage{
howard37
}
\description{
Dataset containing statisics for plywood products(3/8-in. basis).
All units are in Million Square ft. unless otherwise specified.
}
\keyword{howard37}

