% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{howard55}
\alias{howard55}
\title{Insulating board production, imports, exports, and consumption, 1965-1999.}
\format{A data frame with 56 observations of  5 variables:
\describe{
\item{Production}{Insulating board production}
\item{Imports}{Insulating board imports}
\item{Exports}{Insulating board exports}
\item{Total.Consumption}{Insulating board total}
\item{PerCapita.Consumption}{Per Capita Consumption in Pounds}
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard55
}
\description{
All units in thousand tons except when specified otherwise.
}
\keyword{datasets}

