% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocSami.R
\docType{data}
\name{howard56}
\alias{howard56}
\title{Hardboard production, imports, exports and consumption, 1965-1999.}
\format{A data frame with 56 observations on 5 variables. (all units are in million square feet except when specified otherwise)
\itemize{
\item \code{Production} (Production)
\item \code{Imports}(Imports)
\item \code{Exports}(Exports)
\item \code{Total}(Total)
\item \code{PerCapita}(Per Capita units: Square feet)
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard56
}
\description{
Hardboard production, imports, exports and consumption, 1965-1999.
}
\keyword{howard56}

