% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{howard7}
\alias{howard7}
\title{Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.}
\format{A data frame with 56 observations on 31 variables. (all units are in million cubic feet except when specified otherwise)
\itemize{
\item \code{AllProduction.Prod} (All products production)
\item \code{AllProduct.Consump}(All products consumption)
\item \code{Indu.RW.Tot.Prod}(Industrial roundwood uses total production)
\item \code{Indu.RW.Tot.Imports}(Industrial roundwood uses total imports)
\item \code{Indu.RW.Tot.Exports}(Industrial roundwood uses total exports)
\item \code{Indu.RW.Tot.Consump}(Industrial roundwood uses total consumption)
\item \code{Indu.RW.Lum.Prod}(Industrial roundwood uses lumber production)
\item \code{Indu.RW.Lum.Imports}(Industrial roundwood uses lumber imports)
\item \code{Indu.RW.Lum.Exports}(Industrial roundwood uses lumber exports)
\item \code{Indu.RW.Lum.Consump}(Industrial roundwood uses lumber consumption)
\item \code{Indu.RW.PlyandVen.Prod}(Industrial roundwood uses plywood and veneer production)
\item \code{Indu.RW.PlyamdVen.Imports}(Industrial roundwood uses plywood and veneer imports)
\item \code{Indu.RW.PlyandVen.Exports}(Industrial roundwood uses plywood and veneer exports)
\item \code{Indu.RW.PlyandVen.Consump}(Industrial roundwood uses plywood and veneer consumption)
\item \code{Indu.RW.Pulp.Prod}(Industrial roundwood uses pulp-based products production)
\item \code{Indu.RW.Pulp.Imports}(Industrial roundwood uses pulp-based products imports)
\item \code{Indu.RW.Pulp.Exports}(Industrial roundwood uses pulp-based products exports)
\item \code{Indu.RW.Pulp.Consump}(Industrial roundwood uses pulp-based products consumption)
\item \code{Indu.RW.OtherIndustrial.ProdAndConsump}(Other industrial product production and consumption)
\item \code{Indu.RW.Logs.Imports}(Industrial roundwood uses log imports)
\item \code{Indu.RW.Logs.Exports}(Industrial roundwood uses log exports)
\item \code{Indu.RW.Pulp.Imports}(Industrial roundwood uses pulpwood chip imports)
\item \code{Indu.RW.Pulp.Exports}(Industrial roundwood uses pulpwood chip exports)
\item \code{FuelWood.ProdAndConsumption}(Fuelwood production and consumption)
\item \code{UnNamed1}
\item \code{UnNamed2}
\item \code{UnNamed3}
\item \code{UnNamed4}
\item \code{UnNamed5}
\item \code{UnNamed6}
}}
\source{
USFS Estimations and Bureau of Census Data
}
\usage{
howard7
}
\description{
Production, imports, exports, and consumption of softwood products, by major product, 1965-1999.
}
\keyword{howard7}

