% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocSami.R
\docType{data}
\name{ince1}
\alias{ince1}
\title{U.S. Annual industrial wood product in conventional units of measure.
 :(units are in Million feet squared 3/8in.):}
\format{A data frame with 151 observations of  25 variables
\itemize{
\item \strong{SW.Ply} (Softwood Plywood units: million ft^2, 3/8 in)
\item \strong{OSB.Wafer.board} (OSB/ Wafer board units: million ft^2, 3/8 in)
\item \strong{Lam.Ven.Lumb}(Laminated Veneer lumber units: million ft^3)
\item \strong{HW.Ply.Ven}(Hardwood, Plywood and veneer units: million ft^2, 3/8 in)
\item \strong{SW.Lumb}(Softwood lumber units: million bd. ft.)
\item \strong{HW.Lumb}(Hardwood lumber units: million bd. ft.)
\item \strong{Lumb.Pallet.Plant}(Lumber made at Pallet plants units: million bd. ft.)
\item \strong{PartBoard.Prod}(Particalboard production units: million ft^2, 3/4 in.)
\item \strong{Hardboard.Prod}(Hardboard production units: million ft^2, 1/8 in.)
\item \strong{MDF.Pro}(MDF production units: million ft^2, 3/4 in.)
\item \strong{Pulp.Paper.Board}(Pulp, paper and board production units: Thousand tons)
\item \strong{Other.Products}(Other industrial products units: million ft^3)
\item \strong{InsulatingBoard}(Insulating board units: million ft^2, 1/2 in.)
\item \strong{Fuelwood}(Fuelwood units: million ft^3)
\item \strong{FuelwoodT}(Fuelwood units: Thousand short tons)
\item \strong{Log.Chip.Exports}(Log and Chip exports units: Thousand short tons)
\item \strong{RW.Dom.Prod}(Roundwood Equivalent of Domestic Product Output units: Thousand short tons)
\item \strong{Estimated.Tot.Dom.Timber}(Estimated Total Domestic Timber Harvest units: Thousand short tons)
}}
\source{
Sources listed on the excel spread from the National Forest Services
}
\usage{
ince1
}
\description{
U.S. Annual industrial wood product in conventional units of measure.
 :(units are in Million feet squared 3/8in.):
}
\keyword{ince1}

