% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{IncePaper}
\alias{IncePaper}
\title{U.S. Paper & Paperboard, Production, Imports, Exports, Apparent Consumption, 1900 - 2050
Data set containing statistics for paper and paperboard in the U.S.}
\format{\preformatted{'data.frame':	152 obs. of  10 variables:
 $ Paper.board.prod                : num  2168 2355 2543 2731 2919 ...
 $ Paper.board.imports             : num  5 10 15 20 25 30 35 40 45 50 ...
 $ Paper.board.ApparentConsumption : num  75 75 75 75 75 ...
 $ Population                      : num  2098 2290 2483 2676 2869 ...
 $ Paper.board.ConsumptionPerCapita: num  74900 76353 77806 79259 80712 ...
 $ NA                              : num  56 60 63.8 67.5 71.1 ...
 $ NA                              : Factor w/ 15 levels "78299","80477",..: NA NA NA NA NA NA NA NA NA NA ...
 $ NA                              : Factor w/ 14 levels "12507","13057",..: NA NA NA NA NA NA NA NA NA NA ...
 $ NA                              : Factor w/ 18 levels "11456","11506",..: 15 17 14 18 NA NA NA NA NA NA ...
 $ NA                              : Factor w/ 6 levels "13745","14501",..: NA NA NA NA NA NA NA NA NA NA ...
}}
\usage{
IncePaper
}
\description{
U.S. Paper & Paperboard, Production, Imports, Exports, Apparent Consumption, 1900 - 2050
Data set containing statistics for paper and paperboard in the U.S.
}
\keyword{datasets}

