% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lumberpre1900}
\alias{lumberpre1900}
\title{Lumber wood in single and multifamily houses build 1800 to 1900}
\format{A vector with 151 observations of 1 variable:
\describe{
\item{Carbon.Lumberwood}{Carbon from lumber wood in single and multifamily houses in od tons wood fiber}
}}
\source{
'ftp://ftp2.fs.fed.us/incoming/fpl/Skog/[Woodcarb II special lumber July 5 2007 Base Case.xls]SW Calc'!$V3878
}
\usage{
lumberpre1900
}
\description{
Lumber wood in single and multifamily houses build 1800 to 1900
}
\keyword{datasets}

