% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalVariables.R
\name{swp_carbon_stockchange}
\alias{swp_carbon_stockchange}
\title{Calculates changes in carbon stock from solid wood products.
Units are in Tg Carbon.
For production approach, values correspond to `Calculation$BY` in the spreadsheet.
For stock change approach, values correspond to `Calculation$H` in the spreadsheet.}
\usage{
swp_carbon_stockchange(years, approach = c("Production", "Stock Change"),
  decaydistribution = c("Exponential", "K=2", "K=10"),
  halflives = halfLives, fsp = fracstrpanels, fnsp = fracnonstrpanels,
  fsawn = fracsawnwood, swpdata)
}
\arguments{
\item{years}{years to return carbon totals for}

\item{approach}{approach type used to calculate}

\item{decaydistribution}{type of decay distribution to use}

\item{halflives}{data frame of half lives to use. Must have a column associated with each
one of 13 end uses and half life values for 1900 until the maximum year of interest.}

\item{fsp}{Fraction of structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fnsp}{Fraction of non-structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fsawn}{Fraction of sawnwood products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{swpdata}{Optional argument to provide swp production and trade time series
for purposes of error analysis}
}
\value{
A vector of swp carbon stock changes for `years`
}
\description{
Calculates changes in carbon stock from solid wood products.
Units are in Tg Carbon.
For production approach, values correspond to `Calculation$BY` in the spreadsheet.
For stock change approach, values correspond to `Calculation$H` in the spreadsheet.
}
\examples{
\dontrun{
swp_carbon_stockchange(1990:2000)
swp_carbon_stockchange(1950:1975, approach = "Stock Change", decaydistribution = "K=2")
}
}

