% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcarbon.R
\name{swpcarbontotal}
\alias{swpcarbontotal}
\title{Total carbon in solid-wood products (od tons of wood fiber)}
\usage{
swpcarbontotal(Yrs = 1990:2015, distribution = c("Exponential", "K=2"),
  onlytotal = TRUE, lumberpre = TRUE, approach = c("Production",
  "Stock Change"), halflives = halfLives)
}
\arguments{
\item{Yrs}{years to calculate carbon totals for}

\item{distribution}{type of decay distribution}

\item{onlytotal}{return total carbon or totals by end uses}

\item{lumberpre}{whether to include carbon from lumber before 1900 in totals}

\item{approach}{approach type to use to calculate}

\item{halflives}{data frame of half lives for end uses. Default half lives are used but
a data frame with 13 columns with half lives for the appropriate years can be used.}
}
\value{
returns either only total for all end uses or total by enduses with option of lumberpre1900
}
\description{
Calculates carbon per end use and total carbon in solidwood products.
}
\details{
ALl units returned are in od tons of wood fiber.
If "onlytotal" is false, total carbon is returned for each one of the 13 end uses.
If "onlytotal" is true, the totals are summed up and one total is returned. test

This corresponds to the `SWCalcP$H4389` column in the WOODCARB spreadsheet for Production approach.
This corresponds to the `SWCalc$H4389` column in the WOODCARB spreadsheet for Stock Change approach.
}
\examples{
swpcarbontotal(onlytotal=FALSE)
swpcarbontotal(1950:1975, approach = "Stock Change")
}

