% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcarbon.R
\name{swpcarbontotal}
\alias{swpcarbontotal}
\title{Calculates total carbon left in Yrs}
\usage{
swpcarbontotal(Yrs = 1990:2015, distribution = c("Exponential", "K=2"),
  onlytotal = TRUE, lumberpre = TRUE, approach = c("Production",
  "Stock Change"), halflives = halfLives)
}
\arguments{
\item{Yrs}{years to calculate carbon totals for}

\item{distribution}{type of decay distribution}

\item{onlytotal}{return total carbon or totals by end uses}

\item{lumberpre}{whether to include carbon from lumber before 1900 in totals}

\item{approach}{approach type to use to calculate}

\item{halflives}{data frame of half lives for end uses}
}
\value{
returns either only total for all end uses or total by enduses with option of lumberpre1900
}
\description{
Calculates carbon per end use and total carbon in solidwood products.
}
\examples{
swpcarbontotal(onlytotal=FALSE)
swpcarbontotal(1950:1975, approach = "Stock Change")
}

