% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalcarbon.R
\name{swpcarbontotal}
\alias{swpcarbontotal}
\title{Calculate total carbon in solidwood products.}
\usage{
swpcarbontotal(Yrs = 1990:2015, decaydistribution = c("Exponential", "K=2",
  "K=10"), onlytotal = TRUE, lumberpre = TRUE, approach = c("Production",
  "Stock Change"), halflives = halfLives, fsp = fracstrpanels,
  fnsp = fracnonstrpanels, fsawn = fracsawnwood)
}
\arguments{
\item{Yrs}{years to calculate totals for.}

\item{decaydistribution}{type of decay distribution.}

\item{onlytotal}{if false, total carbon is returned for each one of the 13 end uses.
if true, one total is summed and returned}

\item{lumberpre}{if false, carbon from lumber before 1900 is excluded from totals}

\item{approach}{approach type to use to calculate.
If onlytotal is false then the data returned won't correspond to data in WOODCARB II
If onlytotal is true:
For production approach, the data returned corresponds to the `SWCalcP$H4389` column in WOODCARB II.
For stock change approach, the data returned corresponds to the `SWCalc$H4389` column in WOODCARB II.}

\item{halflives}{data frame of half lives for end uses. Default half lives are used but
a data frame with 13 columns with half lives for the appropriate years can be used.}

\item{fsp}{Fraction of structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fnsp}{Fraction of non-structural panel products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}

\item{fsawn}{Fraction of sawnwood products that go to each end use.
Default can be substituted with data with a column for each end use and row for years
from 1900 to latest year of interest.}
}
\value{
Output is either a matrix with 13 columns or 1 column based on the onlytotal argument
}
\description{
Calculates total carbon and carbon by end use in solidwood products. Capable of
calculating carbon totals with different approaches, halflives, and decay distributions.
}
\details{
All units returned are in od tons of wood fiber.
}
\examples{
swpcarbontotal(c(1990,2000,2010), onlytotal=FALSE)
swpcarbontotal(1950:1960, approach = "Stock Change")
}

