% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ulrich36}
\alias{ulrich36}
\title{Plywood production, imports, exports, and consumption, by softwoods and hardwoods, 1950-87}
\format{A data frame with 38 observations of 15 variables:
\describe{
\item{Years}{Years}
\item{Prod.Tot}{Total Plywood Production}
\item{Prod.SW}{Softwood Plywood Production}
\item{Prod.HW}{Hardwood Plywood Production}
\item{Imports.Tot}{Total Plywood Imports}
\item{Imports.SW}{Softwood Plywood Imports}
\item{Imports.HW}{Hardwood Plywood Imports}
\item{Exports.Tot}{Total Plywood Exports}
\item{Exports.SW}{Softwood Plywood Exports}
\item{Exports.HW}{Hardwood Plywood Exports}
\item{Consump.Tot}{Total Plywood Consumption}
\item{Consump.SW}{Softwood Plywood Consumption}
\item{Consump.HW}{Hardwood Plywood Consumption}
\item{PerCapitaConsump.Tot}{Total Plywood Consumption Per Capita, in Square ft.}
\item{PerCapitaConsump.SW}{Softwood Plywood Consumption Per Capita, in Square ft.}
\item{PerCapitaConsump.HW}{Hardwood Plywood Consumption Per Capita, in Square ft.}
}
U.S. Department of Commerce, Bureauot Census (40,50.51,52)
American Plywood Association (5).}
\usage{
ulrich36
}
\description{
Dataset containing various statistics for plywood products. (3/8-inch basis)
All units are in Million Square ft unless otherwise specified.
Year 1987 includes small volumes of mixed species (not classified as softwoods or hardwoods).
}
\keyword{datasets}

