% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocSami.R
\docType{data}
\name{ulrich52}
\alias{ulrich52}
\title{Particleboard and medium density fiberboard production, imports, exports, and consumption, 1950-87}
\format{A data frame with 38 observations of  7 variables. (all units in millions square feet except when specified otherwise):
\itemize{
\item \code{Prod.tot} (Production Total)
\item \code{Prod.Part.Board}(Production of Partical Board)
\item \code{Prod.Med.Fiberboard}(Medium Density Fiberboard)
\item \code{Imports}(Imports)
\item \code{Exports}(Exports)
\item \code{Consump.Tot}(Consumption Total)
\item \code{Consump.PerCapita}(Consumption Per Capits (square feet))
}}
\source{
U.S. Department of Commerce, Bureau of the Census
}
\usage{
ulrich52
}
\description{
Particleboard and medium density fiberboard production, imports, exports, and consumption, 1950-87
}
\keyword{ulrich52}

