% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{ulrich6}
\alias{ulrich6}
\title{Production, imports, exports, and consumption of hardwood timber products, by major product, 1950-87}
\format{A data frame with 48 observations of 22 variables.}
\source{
Based on data published by the U.S. Departments of Commerce and Agriculture.
}
\usage{
ulrich6
}
\description{
Dataset containing statistics for hardwood timber products.
All units are in million cubic ft.
}
\details{
\itemize{
\item \strong{AllProduction.Prod} Total production of all products
\item \strong{AllProduct.Consump} Total consumption of all products
\item \strong{Indu.RW.Tot.Prod} Total production of all industrial roundwood products
\item \strong{Indu.RW.Tot.Imports} Total imports of all industrial roundwood products
\item \strong{Indu.RW.Tot.Exports} Total exports of all industrial roundwood products
\item \strong{Indu.RW.Tot.Consump} Total consumption of all industrial roundwood products
\item \strong{Indu.RW.Lum.Prod} Production of lumber products
\item \strong{Indu.RW.Lum.Imports} Imports of lumber products
\item \strong{Indu.RW.Lum.Exports} Exports of lumber products
\item \strong{Indu.RW.Lum.Consump} Consumption of lumber products
\item \strong{Indu.RW.PlyandVen.Prod} Production of plywood and veneer products
\item \strong{Indu.RW.PlyandVen.Imports} Imports of plywood and veneer products
\item \strong{Indu.RW.PlyandVen.Exports} Exports of plywood and veneer products
\item \strong{Indu.RW.PlyandVen.Consump} Consumption of plywood and veneer products
\item \strong{Indu.RW.Pulp.Prod} Production of pulp products
\item \strong{Indu.RW.Pulp.Imports} Imports of pulp products
\item \strong{Indu.RW.Pulp.Exports} Exports of pulp products
\item \strong{Indu.RW.Pulp.Consump} Consumption of pulp products
\item \strong{Indu.RW.OtherIndustrial.ProdAndConsump} Production and consumption of Other industrial roundwood products
\item \strong{Indu.RW.Logs.Imports} Imports of logs
\item \strong{Indu.RW.Logs.Exports} Exports of logs
\item \strong{FuelWood.ProdAndConsumption} Production and Consumption of Fuelwood
\item \strong{UnNamed1} Unnamed
}
}
\keyword{datasets}

