% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{woodToDumps}
\alias{woodToDumps}
\title{Percent of wood to dumps}
\format{A data frome with 151 observations of 1 variable:

\describe{
\item{WoodToDumps}{Percentage of wood that goes to dumps}
}}
\source{
Linked spreadsheet (check this)
}
\usage{
woodToDumps
}
\description{
Gives percentage of all wood that goes to dumps.
}
\keyword{datasets}

