% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.r
\docType{data}
\name{woodToLandFills}
\alias{woodToLandFills}
\title{Percentage of wood that goes to landfills.}
\format{A data frame with 151 observations of 1 variable

\itemize{
\strong{WoodToLandFills} Percentage of wood going to landfills
}}
\source{
Linked spreacsheet (check this)
}
\usage{
woodToLandFills
}
\description{
Units are percentages and the vector is numeric.
}
\keyword{woodToLandFills}

