% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{woodToLandFills}
\alias{woodToLandFills}
\title{Percentage of wood to landfills.
Corresponds to colum U20 in Param + Results Sheet.
Gives percentage of all wood that goes to landfills.}
\format{A data frame with 151 observations of 1 variable:
\describe{
\item{Years}{Years}
\item{WoodToLandFills}{Percentage of wood going to landfills}
}}
\source{
Linked spreadsheet (check this)
}
\usage{
woodToLandFills
}
\description{
Percentage of wood to landfills.
Corresponds to colum U20 in Param + Results Sheet.
Gives percentage of all wood that goes to landfills.
}
\keyword{datasets}

